+++ src/main/java/study/step5/RacingApplication.java
@@ -1,6 +1,6 @@
 package study.step5;
 
-import study.step5.ui.InputView;
+import study.step5.view.InputView;
 
 public class RacingApplication {
     public static void main(String[] args) {
