+++ src/test/java/study/step5/CarTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import study.step5.model.Car;
+import study.step5.domain.Car;
 
 /**
  * Car class Test
