+++ src/test/java/study/step5/RacingWinnerTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import study.step5.model.Car;
+import study.step5.domain.Car;
 
 public class RacingWinnerTest {
     Car car1;
